requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1;python_version<"3.9"
Django==4.1.4
django-ninja==0.20.0
psycopg2-binary
pydantic==1.10.2
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2022.7
